+++ R/metadata.R
@@ -1,5 +1,7 @@
 do_plotting_metadata <- function(iso3) {
   metadata <- naomi::get_plotting_metadata(iso3)
+  metadata <- metadata[metadata$data_type %in%
+                         c("survey", "anc", "output", "programme"), ]
   metadata <- metadata[order(metadata$indicator_sort_order), ]
   lapply(split(metadata, metadata$data_type), build_data_type_metadata)
 }
@@ -30,15 +32,19 @@ build_indicator_metadata <- function(metadata) {
     colour = scalar(metadata$colour),
     invert_scale = scalar(metadata$invert_scale),
     scale = scalar(metadata$scale),
-    accuracy = scalar(metadata$accuracy),
+    accuracy = if (is.na(metadata$accuracy)) {
+      json_null()
+    } else {
+      scalar(metadata$accuracy)
+    },
     format = scalar(metadata$format)
   )
 }
 
-get_barchart_metadata <- function(output) {
+get_barchart_metadata <- function(output, data_type = "output") {
   metadata <- naomi::get_metadata()
   metadata <- metadata[
-    metadata$data_type == "output" & metadata$plot_type == "barchart",
+    metadata$data_type == data_type & metadata$plot_type == "barchart",
     c("indicator", "value_column", "error_low_column", "error_high_column",
       "indicator_column", "indicator_value", "indicator_sort_order",
       "name", "scale", "accuracy", "format")]