+++ R/utils.R
@@ -53,6 +53,10 @@ vcapply <- function(X, FUN, ...) {
   vapply(X, FUN, ..., FUN.VALUE = character(1))
 }
 
+vlapply <- function(X, FUN, ...) {
+  vapply(X, FUN, ..., FUN.VALUE = logical(1))
+}
+
 # This utility will not be needed once all endpoints use file objects
 # rather than paths, at which point this could be moved into the test
 # helpers.