+++ tests/testthat/test-endpoints-chart-data.R
@@ -8,7 +8,7 @@ test_that("input time series works with programme/art data", {
   expect_equal(names(out), c("data", "metadata"))
   expect_true(nrow(out$data) > 100)
   expect_equal(names(out$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "time_step"))
+               c("plot_type", "area_level", "quarter"))
 })
 
 test_that("input time series works with anc data", {
@@ -21,7 +21,7 @@ test_that("input time series works with anc data", {
   expect_equal(names(out), c("data", "metadata"))
   expect_true(nrow(out$data) > 100)
   expect_equal(names(out$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "age"))
+               c("plot_type", "area_level", "age", "quarter"))
 })
 
 test_that("input time series works if both anc and programme are provided", {
@@ -42,7 +42,7 @@ test_that("input time series works if both anc and programme are provided", {
   expect_equal(names(out), c("data", "metadata"))
   expect_true(nrow(out$data) > 100)
   expect_equal(names(out$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "age"))
+               c("plot_type", "area_level", "age", "quarter"))
 })
 
 test_that("input_time_series throws error if unknown file type", {
@@ -78,7 +78,24 @@ test_that("endpoint_input_time_series_plot works with programme data", {
   expect_equal(names(body$data), c("data", "metadata"))
   expect_true(nrow(body$data$data) > 100)
   expect_equal(names(body$data$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "time_step"))
+               c("plot_type", "area_level", "quarter"))
+})
+
+test_that("endpoint_input_time_series_plot works without vls indicators", {
+  endpoint <- endpoint_input_time_series_plot()
+  input <- input_time_series_request(
+    file.path("testdata", "programme_no_vls.csv"),
+    "programme",
+    file.path("testdata", "malawi.geojson"))
+  res <- endpoint$run("programme", input)
+  expect_equal(res$status_code, 200)
+  body <- jsonlite::fromJSON(res$body)
+  expect_equal(body$status, "success")
+  expect_null(body$errors)
+  expect_equal(names(body$data), c("data", "metadata"))
+  expect_true(nrow(body$data$data) > 100)
+  expect_equal(names(body$data$metadata$defaults$selected_filter_options),
+               c("plot_type", "area_level", "quarter"))
 })
 
 test_that("endpoint_input_time_series_plot works with anc data", {
@@ -95,7 +112,7 @@ test_that("endpoint_input_time_series_plot works with anc data", {
   expect_equal(names(body$data), c("data", "metadata"))
   expect_true(nrow(body$data$data) > 100)
   expect_equal(names(body$data$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "age"))
+               c("plot_type", "area_level", "age", "quarter"))
 })
 
 test_that("api can return input time series data for programme/art", {
@@ -115,7 +132,7 @@ test_that("api can return input time series data for programme/art", {
   expect_equal(names(body$data), c("data", "metadata"))
   expect_true(nrow(body$data$data) > 100)
   expect_equal(names(body$data$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "time_step"))
+               c("plot_type", "area_level", "quarter"))
 })
 
 test_that("api can return input time series data for anc", {
@@ -135,5 +152,5 @@ test_that("api can return input time series data for anc", {
   expect_equal(names(body$data), c("data", "metadata"))
   expect_true(nrow(body$data$data) > 100)
   expect_equal(names(body$data$metadata$defaults$selected_filter_options),
-               c("plot_type", "area_level", "age"))
+               c("plot_type", "area_level", "age", "quarter"))
 })